requirements.txt
requests
pillow
pygal     
cairosvg   